% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependentGP-package.R
\docType{package}
\name{dependentGP}
\alias{dependentGP}
\alias{dependentGP-package}
\title{dependentGP}
\description{
\code{dependentGP} is a package for dependent Gaussian processes.
}
\author{
Samuel I. Berchuck \email{sib2@duke.edu}
}
