% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Utilities.R
\name{is.lmcGP}
\alias{is.lmcGP}
\title{is.lmcGP}
\usage{
is.lmcGP(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
\code{is.lmcGP} is a general test of an object being interpretable as a
\code{\link{lmcGP}} object.
}
\details{
The \code{\link{lmcGP}} class is defined as the regression object that
 results from the \code{\link{lmcGP}} regression function.
}
