% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRED_predict.R
\name{predict.lmcGP}
\alias{predict.lmcGP}
\title{predict.lmcGP}
\usage{
\method{predict}{lmcGP}(object, NewTimes, type = "response",
  Verbose = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{lmcGP}} model object for which predictions
are desired from.}

\item{NewTimes}{a numeric vector including desired time(s) points for prediction.}

\item{type}{a character string indicating the type of predictions to be returned. Either, 
"response" or "derivative".}

\item{Verbose}{A logical indicating whether MCMC sampler progress should be printed.}

\item{...}{other arguments.}
}
\value{
\code{predict.lmcGP} returns a \code{list} for type = "response" and a \code{matrix} for
 type = "derivative".

  \describe{

  \item{type = "response"}{For type = "response", the list contains the predictive samples for gamma, theta, 
  and the observed data y. Each of these are arrays with three dimensions. The first dimension
  is the number of time points to be predicted, the second the number of latent dimensions, and
  finally, the third dimensions is the number of posterior samples specified in the original model.}

  \item{type = "derivative"}{An array of the derivative samples. The array has three dimensions. The first dimension
  is the number of time points to be predicted, the second the number of latent dimensions, and
  finally, the third dimensions is the number of posterior samples specified in the original model.}

  }
}
\description{
Predicts future observations from the \code{\link{lmcGP}} model.
}
\details{
\code{predict.lmcGP} uses Bayesian krigging to predict the various processes.
}
\author{
Samuel I. Berchuck
}
